--- a/Cargo.toml
+++ b/Cargo.toml
@@ -57,6 +57,7 @@ env_logger = "0.10.0"
 stripmargin = "0.1.1"
 convert_case = "0.6.0"
 ring = "0.17.5"
+num_cpus = "1.16.0"
 
 [dev-dependencies]
 criterion = "0.5.1"