--- a/examples/.tailcallrc.graphql
+++ b/examples/.tailcallrc.graphql
@@ -6,7 +6,7 @@ directive @server(
   enableQueryValidation: Boolean
   enableResponseValidation: Boolean
   globalResponseTimeout: Int
-  worker: Int
+  workers: Int
   port: Int
   vars: [KeyValue]
   responseHeaders: [KeyValue]