--- a/src/http/server.rs
+++ b/src/http/server.rs
@@ -79,22 +79,19 @@ pub async fn start_server(config: Config) -> Result<()> {
     async move { Ok::<_, anyhow::Error>(service_fn(move |req| handle_request(req, state.clone()))) }
   });
   let addr = (blueprint.server.hostname, blueprint.server.port).into();
-  let rt = tokio::runtime::Builder::new_multi_thread()
-    .worker_threads(blueprint.server.worker)
-    .enable_all()
-    .build()
-    .unwrap();
-
-  tokio::task::spawn_blocking(move || {
-    let _ = rt.block_on(async move {
-      let server = hyper::Server::try_bind(&addr).map_err(CLIError::from)?.serve(make_svc);
-      log::info!("🚀 Tailcall launched at [{}]", addr);
-      if let Some(enable_graphiql) = blueprint.server.enable_graphiql {
-        log::info!("🌍 Playground: http://{}{}", addr, enable_graphiql);
-      }
-      server.await.map_err(CLIError::from)
-    });
-  });
+  let mut builder = &mut tokio::runtime::Builder::new_multi_thread();
+  if blueprint.server.worker != 0 {
+    builder = builder.worker_threads(blueprint.server.worker);
+  }
+  let rt = builder.enable_all().build().unwrap();
+  let _ = rt.spawn(async move {
+    let server = hyper::Server::try_bind(&addr).map_err(CLIError::from)?.serve(make_svc);
+    log::info!("🚀 Tailcall launched at [{}]", addr);
+    if let Some(enable_graphiql) = blueprint.server.enable_graphiql {
+      log::info!("🌍 Playground: http://{}{}", addr, enable_graphiql);
+    }
+    server.await.map_err(CLIError::from)
+  }).await;
 
   Ok(())
 }