**Summary:**  

Control the number of workers on tailcall via a configuration

**Issue Reference(s):**  

/claim #472
fixes #472 

**Build & Testing:**

- [x] I ran `cargo test` successfully.
- [x] I have run `./lint.sh --mode=fix` to fix all linting issues raised by `./lint.sh --mode=check`.