@kremedev 2023-11-11
## Before

|Thread Stats|Avg|Stdev|Max|+/- Stdev|
|---|---|---|---|---|
|Latency|1.59ms|472.89us|22.51ms|98.93%|
|Req/Sec|63.52k|6.28k|76.31k|61.79%|

1902446 requests in 30.10s, 558.80MB read

Non-2xx or 3xx responses: 1902446

Requests/sec:  63196.50

Transfer/sec:     18.56MB

## After

|Thread Stats|Avg|Stdev|Max|+/- Stdev|
|---|---|---|---|---|
|Latency|1.61ms|380.09us|21.42ms|98.57%|
|Req/Sec|62.65k|5.08k|75.43k|65.12%|

1876053 requests in 30.10s, 551.05MB read

Non-2xx or 3xx responses: 1876053

Requests/sec:  62320.30

Transfer/sec:     18.31MB